requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.4b2
buttercms-python==1.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.2
django-environ==0.4.5
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
